refactor JointDataset
